:bug: Fix handling of MongoDB connection / disconnection.
Allow the application to start up without a MongoDB connection, but
expose the resource endpoints nonetheless (the GTG check should prevent
application requests being directed to this service). The resource and
service classes will self-recover when a MongoDB connection is
available.

Periodic monitoring will invoke the isConnected() method in the
MongoDocumentStoreService.
- Whenever the application is newly connected to MongoDB, the desired
index state will be applied. A separate healthcheck reports on the index
state, but this does not influence the GTG response, which is based on
the connection state only. (This allows reindexing to take a long time
without any service management such as k8s considering a pod to be
unresponsive; requests may be served but some may not be performant, if
they rely on absent indexes).
- Whenever the application becomes disconnected from MongoDB, the
application discards its knowledge of the MongoDB index state, so the
process will be invoked again on reconnection.